+++ src/dfx/src/lib/network/id.rs
@@ -8,12 +8,17 @@ use time::OffsetDateTime;
 #[derive(Deserialize, Serialize, Debug)]
 struct NetworkMetadata {
     created: OffsetDateTime,
+    settings_digest: String,
 }
 
 #[context("Failed write network id to {}.", local_server_descriptor.network_id_path().display())]
 pub fn write_network_id(local_server_descriptor: &LocalServerDescriptor) -> DfxResult {
+    let settings_digest = local_server_descriptor
+        .settings_digest()
+        .to_string();
     let contents = NetworkMetadata {
         created: OffsetDateTime::now_utc(),
+        settings_digest,
     };
     let contents =
         serde_json::to_string_pretty(&contents).context("Failed to pretty-format to string")?;