+++ src/dfx/src/actors/replica.rs
@@ -323,6 +323,8 @@ fn replica_start_thread(
             "ecdsa:Secp256k1:dfx_test_key",
             "--chain-key-ids",
             "schnorr:Bip340Secp256k1:dfx_test_key",
+            "--chain-key-ids",
+            "schnorr:Ed25519:dfx_test_key",
             "--log-level",
             &config.log_level.as_ic_starter_string(),
             "--use-specified-ids-allocation-range",