+++ e2e/tests-dfx/start.bash
@@ -87,29 +87,31 @@ teardown() {
   assert_contains "Module hash: None"
 }
 
-
-@test "icx-proxy domain configuration in string form" {
+@test "pocket-ic proxy domain configuration in string form" {
   create_networks_json
   jq '.local.proxy.domain="xyz.domain"' "$E2E_NETWORKS_JSON" | sponge "$E2E_NETWORKS_JSON"
 
   dfx_start
 
-  assert_command ps aux
-  assert_match "icx-proxy.*--domain xyz.domain"
+  domains="$(curl "http://localhost:$(get_pocketic_proxy_config_port)/http_gateway" \
+    | jq -c ".[] | select(.port == $(get_webserver_port)) | .domains | sort")"
+
+  assert_eq '["xyz.domain"]' "$domains"
 }
 
-@test "icx-proxy domain configuration in vector form" {
+@test "pocket-ic proxy domain configuration in vector form" {
   create_networks_json
   jq '.local.proxy.domain=["xyz.domain", "abc.something"]' "$E2E_NETWORKS_JSON" | sponge "$E2E_NETWORKS_JSON"
 
   dfx_start
 
-  assert_command ps aux
-  assert_match "icx-proxy.*--domain xyz.domain"
-  assert_match "icx-proxy.*--domain abc.something"
+  domains="$(curl "http://localhost:$(get_pocketic_proxy_config_port)/http_gateway" \
+    | jq -c ".[] | select(.port == $(get_webserver_port)) | .domains | sort")"
+
+  assert_eq '["abc.something","xyz.domain"]' "$domains"
 }
 
-@test "icx-proxy domain configuration from project defaults" {
+@test "pocket-ic proxy domain configuration from project defaults" {
   dfx_new
   define_project_network
 
@@ -117,17 +119,19 @@ teardown() {
 
   dfx_start
 
-  assert_command ps aux
-  assert_match "icx-proxy.*--domain xyz.domain"
-  assert_match "icx-proxy.*--domain abc.something"
+  domains="$(curl "http://localhost:$(get_pocketic_proxy_config_port)/http_gateway" \
+    | jq -c ".[] | select(.port == $(get_webserver_port)) | .domains | sort")"
+
+  assert_eq '["abc.something","xyz.domain"]' "$domains"
 }
 
-@test "icx-proxy domain configuration from command-line" {
+@test "pocket-ic proxy domain configuration from command-line" {
   dfx_start --domain xyz.domain --domain def.somewhere
 
-  assert_command ps aux
-  assert_match "icx-proxy.*--domain xyz.domain"
-  assert_match "icx-proxy.*--domain def.somewhere"
+  domains="$(curl "http://localhost:$(get_pocketic_proxy_config_port)/http_gateway" \
+    | jq -c ".[] | select(.port == $(get_webserver_port)) | .domains | sort")"
+
+  assert_eq '["def.somewhere","xyz.domain"]' "$domains"
 }
 
 @test "dfx restarts the replica" {
@@ -183,7 +187,7 @@ teardown() {
   assert_command wait_until_replica_healthy
 }
 
-@test "dfx restarts icx-proxy" {
+@test "dfx restarts pocketic proxy" {
   dfx_new_assets hello
   dfx_start
 
@@ -192,23 +196,23 @@ teardown() {
   assert_command dfx canister call hello_backend greet '("Alpha")'
   assert_eq '("Hello, Alpha!")'
 
-  ICX_PROXY_PID=$(get_icx_proxy_pid)
+  POCKETIC_PROXY_PID=$(get_pocketic_proxy_pid)
 
-  echo "icx-proxy pid is $ICX_PROXY_PID"
+  echo "pocket-ic proxy pid is $POCKETIC_PROXY_PID"
 
-  kill -KILL "$ICX_PROXY_PID"
-  assert_process_exits "$ICX_PROXY_PID" 15s
+  kill -KILL "$POCKETIC_PROXY_PID"
+  assert_process_exits "$POCKETIC_PROXY_PID" 15s
 
   ID=$(dfx canister id hello_frontend)
 
   timeout 15s sh -c \
-    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY\"/webserver-port)/sample-asset.txt?canisterId=$ID; do echo waiting for icx-proxy to restart; sleep 1; done" \
-    || (echo "icx-proxy did not restart" && ps aux && exit 1)
+    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY\"/webserver-port)/sample-asset.txt?canisterId=$ID; do echo waiting for pocket-ic proxy to restart; sleep 1; done" \
+    || (echo "pocket-ic proxy did not restart" && ps aux && exit 1)
 
   assert_command curl --fail http://localhost:"$(get_webserver_port)"/sample-asset.txt?canisterId="$ID"
 }
 
-@test "dfx restarts icx-proxy when the replica restarts" {
+@test "dfx restarts pocketic proxy when the replica restarts" {
   [[ "$USE_POCKETIC" ]] && skip "skipped for pocketic: state persistence"
   dfx_new_assets hello
   dfx_start
@@ -219,14 +223,14 @@ teardown() {
   assert_eq '("Hello, Alpha!")'
 
   REPLICA_PID=$(get_replica_pid)
-  ICX_PROXY_PID=$(get_icx_proxy_pid)
+  POCKETIC_PROXY_PID=$(get_pocketic_proxy_pid)
 
   echo "replica pid is $REPLICA_PID"
-  echo "icx-proxy pid is $ICX_PROXY_PID"
+  echo "pocket-ic proxy pid is $POCKETIC_PROXY_PID"
 
   kill -KILL "$REPLICA_PID"
   assert_process_exits "$REPLICA_PID" 15s
-  assert_process_exits "$ICX_PROXY_PID" 15s
+  assert_process_exits "$POCKETIC_PROXY_PID" 15s
 
   timeout 15s sh -c \
     'until dfx ping; do echo waiting for replica to restart; sleep 1; done' \
@@ -252,23 +256,23 @@ teardown() {
   ID=$(dfx canister id hello_frontend)
 
   timeout 15s sh -c \
-    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY/webserver-port\")/sample-asset.txt?canisterId=$ID; do echo waiting for icx-proxy to restart; sleep 1; done" \
-    || (echo "icx-proxy did not restart" && ps aux && exit 1)
+    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY/webserver-port\")/sample-asset.txt?canisterId=$ID; do echo waiting for pocket-ic proxy to restart; sleep 1; done" \
+    || (echo "pocket-ic proxy did not restart" && ps aux && exit 1)
 
   assert_command curl --fail http://localhost:"$(get_webserver_port)"/sample-asset.txt?canisterId="$ID"
 }
 
-@test "dfx restarts icx-proxy when pocketic restarts" {
+@test "dfx restarts pocketic proxy when pocketic restarts" {
   [[ "$USE_POCKETIC" ]] || skip "skipped for replica"
   dfx_start
   POCKETIC_PID=$(get_pocketic_pid)
-  ICX_PROXY_PID=$(get_icx_proxy_pid)
+  POCKETIC_PROXY_PID=$(get_pocketic_proxy_pid)
   echo "pocketic pid is $POCKETIC_PID"
-  echo "icx-proxy pid is $ICX_PROXY_PID"
+  echo "pocketic proxy pid is $POCKETIC_PROXY_PID"
 
   kill -KILL "$POCKETIC_PID"
   assert_process_exits "$POCKETIC_PID" 15s
-  assert_process_exits "$ICX_PROXY_PID" 15s
+  assert_process_exits "$POCKETIC_PROXY_PID" 15s
   
   timeout 15s sh -c \
     'until dfx ping; do echo waiting for replica to restart; sleep 1; done' \
@@ -280,8 +284,8 @@ teardown() {
   ID=$(dfx canister id hello_frontend --network "$POCKETIC_NETWORK")
 
   timeout 15s sh -c \
-    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY/webserver-port\")/sample-asset.txt?canisterId=$ID; do echo waiting for icx-proxy to restart; sleep 1; done" \
-    || (echo "icx-proxy did not restart" && ps aux && exit 1)
+    "until curl --fail http://localhost:\$(cat \"$E2E_SHARED_LOCAL_NETWORK_DATA_DIRECTORY/webserver-port\")/sample-asset.txt?canisterId=$ID; do echo waiting for pocket-ic proxy to restart; sleep 1; done" \
+    || (echo "pocket-ic proxy did not restart" && ps aux && exit 1)
 
   assert_command curl --fail http://localhost:"$(get_webserver_port)"/sample-asset.txt?canisterId="$ID"
 }
@@ -540,5 +544,5 @@ teardown() {
 @test "dfx-started processes can be killed with dfx killall" {
     dfx_start
     dfx killall
-    assert_command_fail pgrep dfx replica icx-proxy
+    assert_command_fail pgrep dfx replica pocket-ic
 }