+++ src/dfx/src/lib/sign/sign_transport.rs
@@ -1,7 +1,7 @@
 use super::signed_message::SignedMessageV1;
 use candid::Principal;
-use ic_agent::agent::Transport;
-use ic_agent::{AgentError, RequestId};
+use ic_agent::agent::{Envelope, Transport};
+use ic_agent::{AgentError, TransportCallResponse};
 use std::fs::{File, OpenOptions};
 use std::future::Future;
 use std::io::{Read, Write};
@@ -82,13 +82,14 @@ impl Transport for SignTransport {
         &'a self,
         _effective_canister_id: Principal,
         envelope: Vec<u8>,
-        request_id: RequestId,
-    ) -> Pin<Box<dyn Future<Output = Result<(), AgentError>> + Send + 'a>> {
+    ) -> Pin<Box<dyn Future<Output = Result<TransportCallResponse, AgentError>> + Send + 'a>> {
         async fn run(
             s: &SignTransport,
             envelope: Vec<u8>,
-            request_id: RequestId,
-        ) -> Result<(), AgentError> {
+        ) -> Result<TransportCallResponse, AgentError> {
+            let request_id = serde_cbor::from_slice::<Envelope>(&envelope)?
+                .content
+                .to_request_id();
             let message = s
                 .message_template
                 .clone()
@@ -110,8 +111,7 @@ impl Transport for SignTransport {
                 .into(),
             ))
         }
-
-        Box::pin(run(self, envelope, request_id))
+        Box::pin(run(self, envelope))
     }
 
     fn query<'a>(