+++ CHANGELOG.md
@@ -11,6 +11,14 @@ The new `dfx canister snapshot` command can be used to create, apply, and delete
 Schnorr signature signing for `Ed25519` is now enabled.
 A test key id `Ed25519:dfx_test_key` is ready to be used by locally created canisters.
 
+## Dependencies
+
+### Replica
+
+Updated replica to elected commit 3d0b3f10417fc6708e8b5d844a0bac5e86f3e17d.
+This incorporates the following executed proposals:
+
+- [131473](https://dashboard.internetcomputer.org/proposal/131473)
 
 # 0.22.1